Change returned type of findAllIdsOfRecipesWithLikedProducts method in Recipe Repository to List of Long objects

+++ src/main/java/yamly/repositories/RecipeRepository.java
@@ -9,5 +9,5 @@ import java.util.List;
 
 public interface RecipeRepository extends JpaRepository<Recipe, Integer> {
     @Query(value = "SELECT recipes_id FROM yamly.products_recipes WHERE products_id IN :liked AND products_id NOT IN :disliked", nativeQuery = true)
-    List<Integer> findAllIdsOfRecipesWithLikedProducts(@Param("liked") List<Integer> liked, @Param("disliked") List<Integer> disliked);
+    List<Integer> findAllIdsOfRecipesWithLikedProducts(@Param("liked") List<Long> liked, @Param("disliked") List<Long> disliked);
 }
